# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for QuokkaQ, rewritten in Go.",
        "title": "QuokkaQ Go Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
...
        "/units/{unitId}/shift/counters": {
            "get": {
                "description": "Retrieves counters for shift view",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Get shift counters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Counter"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/units/{unitId}/shift/dashboard": {
            "get": {
                "description": "Retrieves dashboard statistics for a unit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Get dashboard stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/units/{unitId}/shift/eod": {
            "post": {
                "description": "Performs end of day operations for a unit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Execute End of Day",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/units/{unitId}/shift/queue": {
            "get": {
                "description": "Retrieves current queue tickets for a unit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Get queue tickets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Booking": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "serviceId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                },
                "userPhone": {
                    "type": "string"
                }
            }
        },
        "models.Counter": {
            "type": "object",
            "properties": {
                "assignedTo": {
                    "type": "string"
                },
                "assignedUser": {
                    "$ref": "#/definitions/models.User"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
...
        "models.PreRegistration": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "6-digit unique code",
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "customerPhone": {
                    "type": "string"
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/models.Service"
                },
                "serviceId": {
                    "type": "string"
                },
                "status": {
                    "description": "created, canceled, ticket_issued, completed",
                    "type": "string"
                },
                "ticket": {
                    "$ref": "#/definitions/models.Ticket"
                },
                "ticketId": {
                    "type": "string"
                },
                "time": {
                    "description": "HH:MM",
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "backgroundColor": {
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "description": {
                    "type": "string"
                },
                "descriptionEn": {
                    "type": "string"
                },
                "descriptionRu": {
                    "type": "string"
                },
                "duration": {
                    "description": "In seconds",
                    "type": "integer"
                },
                "gridCol": {
                    "type": "integer"
                },
                "gridColSpan": {
                    "type": "integer"
                },
                "gridRow": {
                    "description": "Grid configuration",
                    "type": "integer"
                },
                "gridRowSpan": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isLeaf": {
                    "type": "boolean"
                },
                "maxWaitingTime": {
                    "description": "In seconds",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nameEn": {
                    "type": "string"
                },
                "nameRu": {
                    "type": "string"
                },
                "numberSequence": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "prebook": {
                    "type": "boolean"
                },
                "prefix": {
                    "type": "string"
                },
                "textColor": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
...
        "models.SlotConfig": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "days": {
                    "description": "[\"monday\", \"tuesday\", ...]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "endTime": {
                    "description": "HH:MM",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interval": {
                    "description": "in minutes",
                    "type": "integer"
                },
                "startTime": {
                    "description": "HH:MM",
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Ticket": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/models.Booking"
                },
                "bookingId": {
                    "type": "string"
                },
                "calledAt": {
                    "type": "string"
                },
                "completedAt": {
                    "type": "string"
                },
                "confirmedAt": {
                    "type": "string"
                },
                "counter": {
                    "$ref": "#/definitions/models.Counter"
                },
                "counterId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isEod": {
                    "type": "boolean"
                },
                "lastCalledAt": {
                    "type": "string"
                },
                "maxWaitingTime": {
                    "description": "Snapshot from Service at creation",
                    "type": "integer"
                },
                "preRegistration": {
                    "description": "No DB FK: avoids AutoMigrate cycle with pre_registrations.ticket_id → tickets.id",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PreRegistration"
                        }
                    ]
                },
                "preRegistrationId": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "queueNumber": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/models.Service"
                },
                "serviceId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "ttsUrl": {
                    "description": "URL to the generated TTS audio file",
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
        "models.Unit": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "companyId": {
                    "type": "string"
                },
                "config": {
                    "type": "object"
                },
                "counters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Counter"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "preRegistrations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PreRegistration"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "slotConfig": {
                    "$ref": "#/definitions/models.SlotConfig"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                },
                "timezone": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserRole"
                    }
                },
                "type": {
                    "type": "string"
                },
                "units": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserUnit"
                    }
                }
            }
        },
        "models.UserRole": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "roleId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.UserUnit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "unit": {
                    "$ref": "#/definitions/models.Unit"
                },
                "unitId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}